refactor: Test refactoring under better file names
